feat: converts to ts `socket-config` file
